+++ app/Providers/AppServiceProvider.php
@@ -24,7 +24,13 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot(): void
     {
-        //
+        
+        Request::setTrustedProxies(
+            ['REMOTE_ADDR'], 
+            Request::HEADER_X_FORWARDED_FOR
+        );
+
+        // For generating API documentation
         if (class_exists(\Knuckles\Scribe\Scribe::class)) {
             Scribe::beforeResponseCall(function (Request $request, ExtractedEndpointData $endpointData) {
                 // Add a header to all requests