+++ app/Providers/AppServiceProvider.php
@@ -24,11 +24,8 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot(): void
     {
-        
-        Request::setTrustedProxies(
-            ['REMOTE_ADDR'], 
-            Request::HEADER_X_FORWARDED_FOR
-        );
+        // Trust the X-Forwarded-For header (for getting client IP address)
+        Request::setTrustedProxies(['REMOTE_ADDR'], Request::HEADER_X_FORWARDED_FOR);
 
         // For generating API documentation
         if (class_exists(\Knuckles\Scribe\Scribe::class)) {