+++ resources/views/emails/partner_email.blade.php
@@ -1,0 +1,4 @@
+@extends('emails.layout')
+@section('content')
+    {!! $content_message !!}
+@endsection